requirements.txt
fastapi[all]
uvicorn
python-multipart
faiss-cpu==1.7.4
numpy<2
